Internalize DD named-gate helpers
