+++ src/dd/Operations.cpp
@@ -13,7 +13,6 @@
 #include "dd/Complex.hpp"
 #include "dd/DDDefinitions.hpp"
 #include "dd/Edge.hpp"
-#include "dd/GateMatrixDefinitions.hpp"
 #include "dd/Package.hpp"
 #include "ir/Definitions.hpp"
 #include "ir/Permutation.hpp"
@@ -37,17 +36,248 @@
 #include <vector>
 
 namespace dd {
+namespace {
 
-MatrixDD getStandardOperationDD(Package& dd, const qc::OpType type,
-                                const std::vector<fp>& params,
-                                const qc::Controls& controls,
-                                const std::vector<qc::Qubit>& targets) {
+GateMatrix uMat(const fp lambda, const fp phi, const fp theta) {
+  return GateMatrix{{{std::cos(theta / 2.), 0.},
+                     {-std::cos(lambda) * std::sin(theta / 2.),
+                      -std::sin(lambda) * std::sin(theta / 2.)},
+                     {std::cos(phi) * std::sin(theta / 2.),
+                      std::sin(phi) * std::sin(theta / 2.)},
+                     {std::cos(lambda + phi) * std::cos(theta / 2.),
+                      std::sin(lambda + phi) * std::cos(theta / 2.)}}};
+}
+
+GateMatrix u2Mat(const fp lambda, const fp phi) {
+  return GateMatrix{
+      SQRT2_2,
+      {-std::cos(lambda) * SQRT2_2, -std::sin(lambda) * SQRT2_2},
+      {std::cos(phi) * SQRT2_2, std::sin(phi) * SQRT2_2},
+      {std::cos(lambda + phi) * SQRT2_2, std::sin(lambda + phi) * SQRT2_2}};
+}
+
+GateMatrix pMat(const fp lambda) {
+  return GateMatrix{1, 0, 0, {std::cos(lambda), std::sin(lambda)}};
+}
+
+GateMatrix rxMat(const fp lambda) {
+  return GateMatrix{{{std::cos(lambda / 2.), 0.},
+                     {0., -std::sin(lambda / 2.)},
+                     {0., -std::sin(lambda / 2.)},
+                     {std::cos(lambda / 2.), 0.}}};
+}
+
+GateMatrix ryMat(const fp lambda) {
+  return GateMatrix{{{std::cos(lambda / 2.), 0.},
+                     {-std::sin(lambda / 2.), 0.},
+                     {std::sin(lambda / 2.), 0.},
+                     {std::cos(lambda / 2.), 0.}}};
+}
+
+GateMatrix rzMat(const fp lambda) {
+  return GateMatrix{{{std::cos(lambda / 2.), -std::sin(lambda / 2.)},
+                     0,
+                     0,
+                     {std::cos(lambda / 2.), std::sin(lambda / 2.)}}};
+}
+
+GateMatrix rMat(const fp theta, const fp phi) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+  const auto sinPhi = std::sin(phi);
+  const auto cosPhi = std::cos(phi);
+  const std::complex<fp> diag = {cosTheta, 0.};
+  const std::complex<fp> m01 = {-sinTheta * sinPhi, -sinTheta * cosPhi};
+  const std::complex<fp> m10 = {sinTheta * sinPhi, -sinTheta * cosPhi};
+
+  return GateMatrix{diag, m01, m10, diag};
+}
+
+TwoQubitGateMatrix rxxMat(const fp theta) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+
+  return TwoQubitGateMatrix{
+      {{cosTheta, 0, 0, {0., -sinTheta}},
+       {0, cosTheta, {0., -sinTheta}, 0},
+       {0, {0., -sinTheta}, cosTheta, 0},
+       {std::complex<fp>{0., -sinTheta}, 0, 0, cosTheta}}};
+}
+
+TwoQubitGateMatrix ryyMat(const fp theta) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+
+  return TwoQubitGateMatrix{{{cosTheta, 0, 0, {0., sinTheta}},
+                             {0, cosTheta, {0., -sinTheta}, 0},
+                             {0, {0., -sinTheta}, cosTheta, 0},
+                             {std::complex<fp>{0., sinTheta}, 0, 0, cosTheta}}};
+}
+
+TwoQubitGateMatrix rzzMat(const fp theta) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+
+  return TwoQubitGateMatrix{{{std::complex<fp>{cosTheta, -sinTheta}, 0, 0, 0},
+                             {0, {cosTheta, sinTheta}, 0, 0},
+                             {0, 0, {cosTheta, sinTheta}, 0},
+                             {0, 0, 0, {cosTheta, -sinTheta}}}};
+}
+
+TwoQubitGateMatrix rzxMat(const fp theta) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+
+  return TwoQubitGateMatrix{{{cosTheta, {0., -sinTheta}, 0, 0},
+                             {std::complex<fp>{0., -sinTheta}, cosTheta, 0, 0},
+                             {0, 0, cosTheta, {0., sinTheta}},
+                             {0, 0, {0., sinTheta}, cosTheta}}};
+}
+
+TwoQubitGateMatrix xxMinusYYMat(const fp theta, const fp beta = 0.) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+  const auto cosBeta = std::cos(beta);
+  const auto sinBeta = std::sin(beta);
+
+  return TwoQubitGateMatrix{
+      {{cosTheta, 0, 0, {-sinBeta * sinTheta, -cosBeta * sinTheta}},
+       {0, 1, 0, 0},
+       {0, 0, 1, 0},
+       {std::complex<fp>{sinBeta * sinTheta, -cosBeta * sinTheta}, 0, 0,
+        cosTheta}}};
+}
+
+TwoQubitGateMatrix xxPlusYYMat(const fp theta, const fp beta = 0.) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+  const auto cosBeta = std::cos(beta);
+  const auto sinBeta = std::sin(beta);
+
+  return TwoQubitGateMatrix{
+      {{1, 0, 0, 0},
+       {0, cosTheta, {sinBeta * sinTheta, -cosBeta * sinTheta}, 0},
+       {0, {-sinBeta * sinTheta, -cosBeta * sinTheta}, cosTheta, 0},
+       {0, 0, 0, 1}}};
+}
+
+GateMatrix singleQubitGateMatrix(const qc::OpType type,
+                                 const std::vector<fp>& params) {
+  switch (type) {
+  case qc::I:
+    return {1, 0, 0, 1};
+  case qc::H:
+    return {SQRT2_2, SQRT2_2, SQRT2_2, -SQRT2_2};
+  case qc::X:
+    return {0, 1, 1, 0};
+  case qc::Y:
+    return {0, {0, -1}, {0, 1}, 0};
+  case qc::Z:
+    return {1, 0, 0, -1};
+  case qc::S:
+    return {1, 0, 0, {0, 1}};
+  case qc::Sdg:
+    return {1, 0, 0, {0, -1}};
+  case qc::T:
+    return {1, 0, 0, {SQRT2_2, SQRT2_2}};
+  case qc::Tdg:
+    return {1, 0, 0, {SQRT2_2, -SQRT2_2}};
+  case qc::SX:
+    return {std::complex<fp>{0.5, 0.5}, std::complex<fp>{0.5, -0.5},
+            std::complex<fp>{0.5, -0.5}, std::complex<fp>{0.5, 0.5}};
+  case qc::SXdg:
+    return {std::complex<fp>{0.5, -0.5}, std::complex<fp>{0.5, 0.5},
+            std::complex<fp>{0.5, 0.5}, std::complex<fp>{0.5, -0.5}};
+  case qc::V:
+    return {SQRT2_2, {0., -SQRT2_2}, {0., -SQRT2_2}, SQRT2_2};
+  case qc::Vdg:
+    return {SQRT2_2, {0., SQRT2_2}, {0., SQRT2_2}, SQRT2_2};
+  case qc::U:
+    return uMat(params.at(2), params.at(1), params.at(0));
+  case qc::U2:
+    return u2Mat(params.at(1), params.at(0));
+  case qc::P:
+    return pMat(params.at(0));
+  case qc::RX:
+    return rxMat(params.at(0));
+  case qc::RY:
+    return ryMat(params.at(0));
+  case qc::RZ:
+    return rzMat(params.at(0));
+  case qc::R:
+    return rMat(params.at(0), params.at(1));
+  default:
+    throw std::invalid_argument("Invalid single-qubit gate type");
+  }
+}
+
+TwoQubitGateMatrix twoQubitGateMatrix(const qc::OpType type,
+                                      const std::vector<fp>& params) {
+  switch (type) {
+  case qc::SWAP:
+    return {{{1, 0, 0, 0}, {0, 0, 1, 0}, {0, 1, 0, 0}, {0, 0, 0, 1}}};
+  case qc::iSWAP:
+    return {{{1, 0, 0, 0}, {0, 0, {0, 1}, 0}, {0, {0, 1}, 0, 0}, {0, 0, 0, 1}}};
+  case qc::iSWAPdg:
+    return {
+        {{1, 0, 0, 0}, {0, 0, {0, -1}, 0}, {0, {0, -1}, 0, 0}, {0, 0, 0, 1}}};
+  case qc::ECR:
+    return {{{0, 0, SQRT2_2, {0, SQRT2_2}},
+             {0, 0, {0, SQRT2_2}, SQRT2_2},
+             {SQRT2_2, {0, -SQRT2_2}, 0, 0},
+             {std::complex<fp>{0., -SQRT2_2}, SQRT2_2, 0, 0}}};
+  case qc::DCX:
+    return {{{1, 0, 0, 0}, {0, 0, 1, 0}, {0, 0, 0, 1}, {0, 1, 0, 0}}};
+  case qc::Peres:
+    return {{{0, 0, 0, 1}, {0, 0, 1, 0}, {1, 0, 0, 0}, {0, 1, 0, 0}}};
+  case qc::Peresdg:
+    return {{{0, 0, 1, 0}, {0, 0, 0, 1}, {0, 1, 0, 0}, {1, 0, 0, 0}}};
+  case qc::RXX:
+    return rxxMat(params.at(0));
+  case qc::RYY:
+    return ryyMat(params.at(0));
+  case qc::RZZ:
+    return rzzMat(params.at(0));
+  case qc::RZX:
+    return rzxMat(params.at(0));
+  case qc::XXminusYY:
+    return xxMinusYYMat(params.at(0), params.at(1));
+  case qc::XXplusYY:
+    return xxPlusYYMat(params.at(0), params.at(1));
+  default:
+    throw std::invalid_argument("Invalid two-qubit gate type");
+  }
+}
+
+ThreeQubitGateMatrix threeQubitGateMatrix(const qc::OpType type) {
+  switch (type) {
+  case qc::RCCX: {
+    ThreeQubitGateMatrix matrix{};
+    for (size_t i = 0; i < THREE_QUBIT_GATE_DIM; ++i) {
+      matrix[i][i] = 1.;
+    }
+    matrix[5][5] = -1.;
+    matrix[6][6] = 0.;
+    matrix[7][7] = 0.;
+    matrix[6][7] = {0., -1.};
+    matrix[7][6] = {0., 1.};
+    return matrix;
+  }
+  default:
+    throw std::invalid_argument("Invalid three-qubit gate type");
+  }
+}
+
+MatrixDD makeStandardOperationDD(Package& dd, const qc::OpType type,
+                                 const std::vector<fp>& params,
+                                 const qc::Controls& controls,
+                                 const std::vector<qc::Qubit>& targets) {
   if (qc::isSingleQubitGate(type)) {
     if (targets.size() != 1) {
       throw std::invalid_argument(
           "Expected exactly one target qubit for single-qubit gate");
     }
-    return dd.makeGateDD(opToSingleQubitGateMatrix(type, params), controls,
+    return dd.makeGateDD(singleQubitGateMatrix(type, params), controls,
                          targets[0U]);
   }
   if (qc::isTwoQubitGate(type)) {
@@ -55,7 +285,7 @@ MatrixDD getStandardOperationDD(Package& dd, const qc::OpType type,
       throw std::invalid_argument(
           "Expected two target qubits for two-qubit gate");
     }
-    return dd.makeTwoQubitGateDD(opToTwoQubitGateMatrix(type, params), controls,
+    return dd.makeTwoQubitGateDD(twoQubitGateMatrix(type, params), controls,
                                  targets[0U], targets[1U]);
   }
   if (qc::isThreeQubitGate(type)) {
@@ -63,22 +293,21 @@ MatrixDD getStandardOperationDD(Package& dd, const qc::OpType type,
       throw std::invalid_argument(
           "Expected three target qubits for three-qubit gate");
     }
-    return dd.makeThreeQubitGateDD(opToThreeQubitGateMatrix(type, params),
-                                   controls, targets[0U], targets[1U],
-                                   targets[2U]);
+    return dd.makeThreeQubitGateDD(threeQubitGateMatrix(type), controls,
+                                   targets[0U], targets[1U], targets[2U]);
   }
   throw std::runtime_error("Unexpected operation type");
 }
 
-MatrixDD getStandardOperationDD(const qc::StandardOperation& op, Package& dd,
-                                const qc::Controls& controls,
-                                const std::vector<qc::Qubit>& targets,
-                                const bool inverse) {
+MatrixDD makeStandardOperationDD(const qc::StandardOperation& op, Package& dd,
+                                 const qc::Controls& controls,
+                                 const std::vector<qc::Qubit>& targets,
+                                 const bool inverse) {
   auto type = op.getType();
 
   if (!inverse) {
-    return getStandardOperationDD(dd, type, op.getParameter(), controls,
-                                  targets);
+    return makeStandardOperationDD(dd, type, op.getParameter(), controls,
+                                   targets);
   }
 
   // invert the operation
@@ -152,8 +381,10 @@ MatrixDD getStandardOperationDD(const qc::StandardOperation& op, Package& dd,
     oss << "negation for gate " << op.getName() << " not available!";
     throw std::runtime_error(oss.str());
   }
-  return getStandardOperationDD(dd, type, params, controls, targetQubits);
+  return makeStandardOperationDD(dd, type, params, controls, targetQubits);
 }
+
+} // namespace
 
 MatrixDD getDD(const qc::Operation& op, Package& dd,
                const qc::Permutation& permutation, const bool inverse) {
@@ -178,7 +409,7 @@ MatrixDD getDD(const qc::Operation& op, Package& dd,
     const auto& targets = permutation.apply(standardOp.getTargets());
     const auto& controls = permutation.apply(standardOp.getControls());
 
-    return getStandardOperationDD(standardOp, dd, controls, targets, inverse);
+    return makeStandardOperationDD(standardOp, dd, controls, targets, inverse);
   }
 
   if (op.isCompoundOperation()) {
